line wrapper to be able to work woth LineData and MyLineData at the same time

+++ stardust/src/main/java/se/de/hu_berlin/informatik/stardust/provider/MyLineData.java
@@ -5,7 +5,7 @@ import net.sourceforge.cobertura.coveragedata.CoverageData;
 public class MyLineData implements Comparable<Object>, CoverageData {
 
 	final private int lineNumber;
-	private int hits;
+	private long hits;
 	
 	public MyLineData(int lineNumber, int hits) {
 		this.lineNumber = lineNumber;
@@ -16,7 +16,7 @@ public class MyLineData implements Comparable<Object>, CoverageData {
 		return lineNumber;
 	}
 	
-	public int getHits() {
+	public long getHits() {
 		return hits;
 	}
